requirements.txt
Django==3.0.8
psycopg2-binary==2.8.4
djangorestframework==3.11.0
django_cron
gunicorn
psycopg2